panel stil now is custom# hook, to be imported as hook from file  PanelComponent, SortingPanelHook-folder
