+++ my-react-app/src/components/Button/Button.component.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Button = ({ onClick, children }) => {
+    return (
+        <button onClick={onClick}>
+            {children}
+        </button>
+    );
+};
+
+export default Button;
