+++ my-react-app/src/components/Button/Button.module.css
@@ -1,0 +1,7 @@
+*{
+    padding: 0;
+    margin: 0;
+    background-color: green;
+    color: #ffff;
+}
+
