+++ my-react-app/src/components/ProductCounter/ProductCounter.jsx
@@ -1,25 +1,23 @@
-import React, { useState } from 'react';
+import React from 'react';
 import classes from './ProductCounter.module.css';
 
-const ProductCounter = () => {
-  const [count, setCount] = useState(0);
+const ProductCounter = ({ count, setCount }) => {
+  const handleInputChange = (e) => {
+      const newValue = e.target.value;
 
-  const increment = () => {
-    setCount(count + 1);
-  };
-
-  const decrement = () => {
-    if (count > 0) {
-      setCount(count - 1);
-    }
+      if (!isNaN(newValue) && newValue >= 0) {
+          setCount(parseInt(newValue)); 
+      }
   };
 
   return (
-    <div className={classes.ProductCounterContainer}>
-      <button onClick={decrement}>-</button>
-      <input type="number" value={count} readOnly />
-      <button onClick={increment}>+</button>
-    </div>
+      <div className={classes.ProductCounterContainer}>
+          <input
+              type="number"
+              value={count}
+              onChange={handleInputChange} 
+          />
+      </div>
   );
 };
