+++ my-react-app/src/components/pages/OneProduct/OneProduct.jsx
@@ -1,12 +1,12 @@
 import React from 'react';
 import OneProductComponent from '../../OneProductComponent/OneProductComponent';
 import classes from './OneProduct.module.css';
-import ProductCounter from '../../ProductCounter/ProductCounter';
+// import ProductCounter from '../../ProductCounter/ProductCounter';
 
 const OneProduct = () => (
   <div className={classes.pageBody}>   
     <OneProductComponent />
-    <ProductCounter />
+    {/* <ProductCounter /> */}
   </div>
 );
