+++ my-react-app/src/index.js
@@ -1,13 +1,13 @@
 import './index.css';
 import App from "./App.jsx";
 import React from 'react';
-import ReactDOM from 'react-dom';
+import ReactDOM from 'react-dom/client';
 import {configureStore} from '@reduxjs/toolkit';
 import { Provider } from "react-redux";
 import rootReducer from './components/SortingFilteringPanel/reducers/index.jsx';
 import initialState from './components/SortingFilteringPanel/store/initialState.jsx';
 import checkboxReducer from './components/SortingFilteringPanel/reducers/checkboxReducer.jsx';
-// import { BrowserRouter } from 'react-router-dom';
+
 
 
 const store = configureStore({
